chore: rename test.md to sample.md for clarity
