+++ apps/web/src/main.ts
@@ -1,6 +1,6 @@
 import './style.css'
 import { createMarkdownRenderer, renderLineChart, renderBarChart, renderTable } from '../../../packages/src';
-import markdownText from '../docs/test.md?raw'
+import markdownText from '../docs/sample.md?raw'
 import { welcomeHtml } from './views/welcomeHtml'
 import { getMarkdownAppHtml } from './views/markdownAppHtml'
 import { showExportModal } from './utils/showExportModal'
